+++ s3pack/obj_list.go
@@ -9,10 +9,12 @@ import (
 	"errors"
 	"fmt"
 	"strings"
+	"sync"
 
 	"github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/aws-sdk-go/aws/awserr"
 	"github.com/aws/aws-sdk-go/service/s3"
+	"github.com/aws/aws-sdk-go/service/s3/s3manager"
 	"github.com/orme292/s3packer/config"
 )
 
@@ -23,6 +25,12 @@ elements of the slice and call the corresponding FileList method. Exceptions are
 See FileList for more information
 */
 type ObjectList []*FileObject
+
+type UploadResult struct {
+	Err         error
+	UploadCount int
+	IgnoreCount int
+}
 
 /*
 NewObjectList is an ObjectList constructor. It takes a slice of paths and returns a slice FileObjects.
@@ -86,7 +94,6 @@ func (objList ObjectList) FixRedundantKeys() error {
 			}
 		}
 	}
-
 	return nil
 }
 
@@ -116,17 +123,17 @@ to retrieve metadata from an S3 object of the same name (s3 key = FileObject.Pre
 the FileObject.Ignore field is set to true and the FileObject.IgnoreString field is set to ErrIgnoreObjectAlreadyExists.
 */
 func (objList ObjectList) IgnoreIfObjectExistsInBucket() {
-	if objList[0].config.Options[config.ProfileOptionOverwrite].(bool) || len(objList) == 0 {
+	if objList[0].c.Options[config.ProfileOptionOverwrite].(bool) || len(objList) == 0 {
 		return
 	}
 
-	sess, _ := NewSession(&objList[0].config)
+	sess, _ := NewSession(&objList[0].c)
 
 	svc := s3.New(sess, &aws.Config{})
 
 	for index := range objList {
 		_, err := svc.HeadObject(&s3.HeadObjectInput{
-			Bucket: aws.String(objList[index].config.Bucket[config.ProfileBucketName].(string)),
+			Bucket: aws.String(objList[index].c.Bucket[config.ProfileBucketName].(string)),
 			Key:    aws.String(objList[index].PrefixedName),
 		})
 		if err != nil {
@@ -188,6 +195,12 @@ func (objList ObjectList) SetChecksum() (err error) {
 	return
 }
 
+func (objList ObjectList) SetGroups() {
+	for index, fo := range objList {
+		fo.SetGroup(index % fo.c.Options[config.ProfileOptionsMaxConcurrent].(int))
+	}
+}
+
 func (objList ObjectList) SetIgnoreIfLocalNotExists() {
 	_ = objList.IterateAndExecute(func(fo *FileObject) (err error) {
 		fo.SetIgnoreIfLocalNotExists()
@@ -212,7 +225,7 @@ See FileObject.SetFileSize for more information.
 */
 func (objList ObjectList) SetFileSizes() {
 	_ = objList.IterateAndExecute(func(fo *FileObject) (err error) {
-		if fo.Ignore {
+		if !fo.Ignore {
 			fo.SetFileSize()
 		}
 		return
@@ -243,16 +256,10 @@ func (objList ObjectList) SetRelativeRoot(dir string) {
 	})
 }
 
-/*
-ReturnNewWithoutIgnored is an ObjectList method. It returns an ObjectList type that contains any FileObject with
-FileObject.Ignore set to false
-*/
-func (objList ObjectList) ReturnNewWithoutIgnored() (newObjList ObjectList) {
+func (objList ObjectList) ReturnTotalUploadedBytes() (total int64) {
 	for index := range objList {
-		if !objList[index].Ignore {
-			newObjList = append(newObjList, objList[index])
-		} else {
-			objList[index].config.Logger.Warn(fmt.Sprintf("Ignoring %s: %s", objList[index].OriginPath, objList[index].IgnoreString))
+		if objList[index].IsUploaded {
+			total += objList[index].FileSize
 		}
 	}
 	return
@@ -277,7 +284,7 @@ See FileObject.Tag for more information.
 */
 func (objList ObjectList) TagOrigins() {
 	_ = objList.IterateAndExecute(func(fo *FileObject) (err error) {
-		if fo.config.Options["tagOrigins"].(bool) {
+		if fo.c.Options["tagOrigins"].(bool) {
 			fo.Tag("Origin", fo.AbsolutePath)
 		}
 		return
@@ -294,7 +301,6 @@ func (objList ObjectList) Upload(c *config.Configuration) (err error, uploaded, 
 		return nil, 0, 0
 	}
 
-	svc, err := BuildUploader(c)
 	if err != nil {
 		return
 	}
@@ -308,13 +314,55 @@ func (objList ObjectList) Upload(c *config.Configuration) (err error, uploaded, 
 	}
 
 	objList.SetIgnoreIfObjExists()
+	objList.SetGroups()
 
-	fi := NewFileIterator(&objList[0].config, objList)
-	if err := svc.UploadWithIterator(aws.BackgroundContext(), fi); err != nil {
-		return err, objList.CountUploaded(), objList.CountIgnored()
+	errs, _, _ := objList.UploadHandler(c)
+	if len(errs) > 0 {
+		for _, err := range errs {
+			c.Logger.Error(fmt.Sprintf("Error during upload: %q", err.Error()))
+		}
+	}
+	return nil, objList.CountUploaded(), objList.CountIgnored()
+}
+
+func (objList ObjectList) UploadHandler(c *config.Configuration) (err []error, uploaded, ignored int) {
+	var wg sync.WaitGroup
+	resultCh := make(chan UploadResult)
+
+	svc, buErr := BuildUploader(c)
+	if buErr != nil {
+		err = append(err, buErr)
+		return err, 0, 0
 	}
 
-	return nil, objList.CountUploaded(), objList.CountIgnored()
+	for i := 0; i < c.Options[config.ProfileOptionsMaxConcurrent].(int); i++ {
+		wg.Add(1)
+		go func(group int, svc *s3manager.Uploader, wg *sync.WaitGroup) {
+			defer wg.Done()
+			fi := NewFileIterator(&objList[0].c, objList, group)
+			err := svc.UploadWithIterator(aws.BackgroundContext(), fi)
+
+			resultCh <- UploadResult{
+				Err:         err,
+				UploadCount: objList.CountUploadedByGroup(group),
+				IgnoreCount: objList.CountIgnoredByGroup(group),
+			}
+		}(i, svc, &wg)
+	}
+
+	go func(wg *sync.WaitGroup) {
+		wg.Wait()
+		close(resultCh)
+	}(&wg)
+
+	for result := range resultCh {
+		if result.Err != nil {
+			err = append(err, result.Err)
+		}
+		uploaded += result.UploadCount
+		ignored += result.IgnoreCount
+	}
+	return
 }
 
 /*
@@ -337,6 +385,15 @@ func (objList ObjectList) CountUploaded() (count int) {
 	return
 }
 
+func (objList ObjectList) CountUploadedByGroup(group int) (count int) {
+	for index := range objList {
+		if objList[index].IsUploaded && objList[index].Group == group {
+			count++
+		}
+	}
+	return
+}
+
 /*
 CountIgnored is an ObjectList method. It returns the number of FileObjects in the ObjectList slice that have the
 FileObject.Ignore field set to true.
@@ -344,6 +401,15 @@ FileObject.Ignore field set to true.
 func (objList ObjectList) CountIgnored() (count int) {
 	for index := range objList {
 		if objList[index].Ignore {
+			count++
+		}
+	}
+	return
+}
+
+func (objList ObjectList) CountIgnoredByGroup(group int) (count int) {
+	for index := range objList {
+		if objList[index].Ignore && objList[index].Group == group {
 			count++
 		}
 	}
@@ -360,8 +426,14 @@ Change this as needed.
 */
 func (objList ObjectList) DebugOutput() {
 	_ = objList.IterateAndExecute(func(fo *FileObject) (err error) {
-		fmt.Println(fo.AbsolutePath)
-		fmt.Println(fo.IsDirectoryPart)
+		fmt.Println()
+		fmt.Printf("AbsolutePath: %q\n", fo.AbsolutePath)
+		fmt.Printf("PrefixedName: %q\n", fo.PrefixedName)
+		fmt.Printf("Group: %d\n", fo.Group)
+		fmt.Printf("IsUploaded: %v\n", fo.IsUploaded)
+		fmt.Printf("Ignored: %v\n", fo.Ignore)
+		fmt.Printf("IgnoreString: %q\n", fo.IgnoreString)
+		fmt.Println()
 		return
 	})
 }