+++ s3pack/types.go
@@ -14,5 +14,5 @@ const (
 )
 
 const (
-	Version = "1.0.1"
+	Version = "1.0.2"
 )