+++ CHANGELOG.md
@@ -3,6 +3,14 @@ Hello Changelog. This is where I document all changes to s3packer. I'll try to r
 or minor tweak. 
 
 ---
+### **1.0.2** (2023-12-13)
+- config: add new options 'maxConcurrentUploads'
+- s3pack: add upload concurrency (handled at ObjectList level)
+- s3pack: config references changed to 'c'
+- s3pack: FileIterator overhaul, group and index tracking used for concurrency
+- s3pack: FileObject has new individual Upload option, but it's unused.
+- s3pack: BucketExists checks are done once before processing any files/dirs (See main.go)
+
 ### **1.0.1** (2023-12-04)
 - use gocritic suggestions
 - resolve gosec scan issues